Split shell.js in a minimal shell.js and extra debug.js part
This way it is easier to disable the debug.js part while including pxscene in the STB builds.
